requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<1.0
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0
django-cors-headers>=4.3,<5.0
